+++ src/main/java/de/alpharogroup/collections/set/SetFactory.java
@@ -89,24 +89,6 @@ public final class SetFactory
 	}
 
 	/**
-	 * Factory method for create new {@link HashSet} and will be returned as {@link Set}
-	 *
-	 * @param <T>
-	 *            the generic type of the elements
-	 * @param initialCapacity
-	 *            the initial capacity
-	 * @return the new {@link HashSet}
-	 * @deprecated use instead the same name method with elements <br>
-	 *             <br>
-	 *             Note: will be removed in next minor release
-	 */
-	@Deprecated
-	public static final <T> Set<T> newHashSet(final int initialCapacity)
-	{
-		return new HashSet<>(initialCapacity);
-	}
-
-	/**
 	 * Factory method for create new {@link LinkedHashSet} and will be returned as {@link Set}
 	 *
 	 * @param <T>
@@ -119,24 +101,6 @@ public final class SetFactory
 	public static final <T> Set<T> newLinkedHashSet(final T... elements)
 	{
 		return newLinkedHashSet(null, elements);
-	}
-
-	/**
-	 * Factory method for create new {@link LinkedHashSet} and will be returned as {@link Set}
-	 *
-	 * @param <T>
-	 *            the generic type of the elements
-	 * @param initialCapacity
-	 *            the initial capacity
-	 * @return the new {@link LinkedHashSet}
-	 * @deprecated use instead the same name method with elements <br>
-	 *             <br>
-	 *             Note: will be removed in next minor release
-	 */
-	@Deprecated
-	public static final <T> Set<T> newLinkedHashSet(final int initialCapacity)
-	{
-		return new LinkedHashSet<>(initialCapacity);
 	}
 
 	/**
@@ -183,24 +147,6 @@ public final class SetFactory
 	public static final <T> Set<T> newInsertionOrderSet(final T... elements)
 	{
 		return newInsertionOrderSet(null, elements);
-	}
-
-	/**
-	 * Factory method for create new {@link InsertionOrderSet} and will be returned as {@link Set}
-	 *
-	 * @param <T>
-	 *            the generic type of the elements
-	 * @param initialCapacity
-	 *            the initial capacity
-	 * @return the new {@link InsertionOrderSet}
-	 * @deprecated use instead the same name method with elements <br>
-	 *             <br>
-	 *             Note: will be removed in next minor release
-	 */
-	@Deprecated
-	public static final <T> Set<T> newInsertionOrderSet(final int initialCapacity)
-	{
-		return new InsertionOrderSet<>(initialCapacity);
 	}
 
 	/**